+++ components/input/index.js
@@ -1,5 +1,7 @@
 import SearchBar from './SearchBar';
+import SignInput from './SignInput';
 
 module.exports = {
   SearchBar,
+  SignInput,
 };
