+++ pages/start/SignIn.jsx
@@ -6,14 +6,11 @@ import {
   View,
   Dimensions,
   TouchableOpacity,
-  TextInput,
-  Form,
 } from 'react-native';
-const diviceWidth = Dimensions.get('window').width;
-
 import { HeaderTitle } from '../../components/header';
-import { SignInput } from '../../components/SignInput';
-import SignUp from './SignUp';
+import { SignInput } from '../../components/input';
+
+const diviceWidth = Dimensions.get('window').width;
 
 export default function SignIn({ navigation }) {
   const [id, setId] = useState('');
@@ -33,11 +30,37 @@ export default function SignIn({ navigation }) {
         MocoMoco 로그인
       </Text>
 
-      <TextInput value={'아이디비밀번호인푹박스나중에'} />
+      <View
+        style={{
+          marginTop: 30,
+          alignItems: 'center',
+        }}
+      >
+        {/* 아이디 */}
+        <SignInput
+          label={'아이디'}
+          value={id}
+          type={'id'}
+          hint={'아이디를 입력하세요'}
+          setValue={setId}
+        />
+
+        {/* 비밀번호 */}
+        <SignInput
+          label={'비밀번호'}
+          value={password}
+          type={'password'}
+          hint={'비밀번호를 입력하세요'}
+          setValue={setPassword}
+        />
+      </View>
 
       {/* 회원가입, 비번찾기 */}
       <View style={{ flexDirection: 'row', margin: 20 }}>
-        <TouchableOpacity style={styles.varifacationTextBox}>
+        <TouchableOpacity
+          style={styles.varifacationTextBox}
+          onPress={() => navigation.navigate('Verification')}
+        >
           <Text style={styles.varifacationText}>회원가입</Text>
         </TouchableOpacity>
         <TouchableOpacity style={styles.passwordTextBox}>
